aios/tools/hape/requirements.txt
protobuf==3.0.0
rsa==4.5
cachetools==2.1.0
kazoo==2.9.0
requests==2.24.0
snakebite==2.11.0
certifi==2021.10.8
selectors2
click==7.1.2
jinja2==2.10.0
markupsafe==0.23
kubernetes==12.0.1
attrs==19.1.0
setuptools==44.1.0
oauthlib==3.1.0
websocket-client==0.59.0